Moving three renderers to another file
